requirements.txt
pyannote.audio
pydub
numpy
structlog
torch
librosa
transformers
openai